+++ HISTORY.rst
@@ -25,6 +25,9 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* Add EAN13 to PO report lines
+* Ported from 7.0 : use refund description in refund wizard
+
 **Bugfixes**
 
 **Build**